pyproject.toml
[project]
name = "glacis"
version = "0.1.0"
description = "Explore the countries that flights arrive from for a specific airport"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "jinja2>=3.1.5",
    "jsonpath-ng>=1.7.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.14.0",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
]
